Refactor useProducts to useSfProducts
