--- packages/data/composables/useProducts/__tests__/useProducts.spec.ts
@@ -1,11 +1,0 @@
-import { useProducts } from '../useProducts';
-
-describe('useProducts', () => {
-  it('should return products', async () => {
-    const { fetchProducts, data } = useProducts();
-
-    await fetchProducts();
-
-    expect(data.value).not.toBeUndefined();
-  });
-});
