+++ packages/data/composables/useSfProducts/__tests__/useProducts.spec.ts
@@ -1,0 +1,11 @@
+import { useSfProducts } from '../useSfProducts';
+
+describe('useSfProducts', () => {
+  it('should return products', async () => {
+    const { fetchProducts, data } = useSfProducts();
+
+    await fetchProducts();
+
+    expect(data.value).not.toBeUndefined();
+  });
+});
